// services/office/src/core/mongo/mongo.enums.ts
export enum MongoErrors {
    Conflict = 'conflict',
    NotFound = 'not-found',
}

export enum ModelIdTag {
    ApiRequest = 'req',
    Business = 'bz',
    User = 'usr',
    Customer = 'cus',
    Webhook = 'wh',
    Account = 'ac',
    Balance = 'bal',
    BalanceHistory = 'bal.h',
    IsvAccount = 'isv.ac',
    SettlementAccount = 'sac',
    SettlementAccountLog = 'sac.log',
    AccountLog = 'ac.log',
    ReserveAccount = 'ac.rsv',
    CardProgram = 'c.prg',
    CardBin = 'c.bin',
    Card = 'c',
    CardTransaction = 'c.txn',
    CardAuthorization = 'c.auth',
    Dispute = 'd',
    Payment = 'p',
    PaymentRequest = 'p.req',
    PaymentReserve = 'p.rsv',
    Event = 'evt',
    EventAttempt = 'evt.a',
    Request = 'req',
    Log = 'log',
    Merchant = 'mch',
    Configuration = 'cfg',
    Fee = 'fee',
    BillProduct = 'bp',
    Billing = 'bil',
    Invoice = 'inv',
    SecretKey = 'sk',
    Approval = 'apr',
    ReconciliationRun = 'rec',
}

export enum ObjectModel {
    User = 'User',
    Customer = 'Customer',
    Webhook = 'Webhook',
    Account = 'Account',
    CardProgram = 'CardProgram',
    Card = 'Card',
    CardTransaction = 'CardTransaction',
    CardAuthorization = 'CardAuthorization',
    Dispute = 'Dispute',
    Payment = 'Payment',
    Event = 'Event',
    EventAttempt = 'EventAttempt',
    Request = 'Request',
    Merchant = 'Merchant',
    Fee = 'Fee',
    Configuration = 'Configuration',
}

export enum ObjectType {
    AccountBalance = 'account.balance',
    SettlementAccountBalance = 'settlement.account.balance',
    CardSecrets = 'card.secrets',
    CardAuthorization = 'card.authorization',
    Payment = 'payment',
    Balance = 'balance',
}